requirements - Copy.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
qrcode[pil]==8.2
python-barcode==0.16.1
Pillow==11.3.0
twilio==9.8.2
python-dotenv==1.1.1